# Conflicts:
#	python/src/com/jetbrains/python/actions/PyRunFileInConsoleAction.java

+++ plugins/InspectionGadgets/InspectionGadgetsAnalysis/src/com/siyeh/ig/fixes/NormalizeDeclarationFix.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2003-2018 Dave Griffith, Bas Leijdekkers
+ * Copyright 2003-2019 Dave Griffith, Bas Leijdekkers
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,59 +18,109 @@ package com.siyeh.ig.fixes;
 import com.intellij.codeInspection.ProblemDescriptor;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.*;
+import com.intellij.psi.impl.source.tree.JavaSharedImplUtil;
+import com.intellij.psi.tree.IElementType;
+import com.intellij.psi.util.PsiTreeUtil;
 import com.siyeh.InspectionGadgetsBundle;
 import com.siyeh.ig.InspectionGadgetsFix;
+import com.siyeh.ig.psiutils.DeclarationSearchUtils;
 import org.jetbrains.annotations.NotNull;
 
 public class NormalizeDeclarationFix extends InspectionGadgetsFix {
 
+  private final boolean myCStyleDeclaration;
+
+  public NormalizeDeclarationFix(boolean cStyleDeclaration) {
+    myCStyleDeclaration = cStyleDeclaration;
+  }
+
   @Override
   @NotNull
   public String getFamilyName() {
-    return InspectionGadgetsBundle.message("normalize.declaration.quickfix");
+    return myCStyleDeclaration
+           ? InspectionGadgetsBundle.message("c.style.array.declaration.replace.quickfix")
+           : InspectionGadgetsBundle.message("normalize.declaration.quickfix");
   }
 
   @Override
   public void doFix(Project project, ProblemDescriptor descriptor) {
-    final PsiElement variableNameElement = descriptor.getPsiElement();
-    final PsiVariable parent =
-      (PsiVariable)variableNameElement.getParent();
-    if (parent == null) {
-      return;
+    PsiElement element = descriptor.getPsiElement();
+    if (!(element instanceof PsiVariable) && !(element instanceof PsiMethod) && !(element instanceof PsiDeclarationStatement)) {
+      element = element.getParent();
     }
-    if (parent instanceof PsiField) {
-      parent.normalizeDeclaration();
-      return;
+    if (element instanceof PsiLocalVariable) {
+      element = element.getParent();
+      if (!(element instanceof PsiDeclarationStatement)) {
+        return;
+      }
     }
-    final PsiElement grandParent = parent.getParent();
-    if (!(grandParent instanceof PsiDeclarationStatement)) {
-      return;
+    if (element instanceof PsiDeclarationStatement) {
+      PsiDeclarationStatement declarationStatement = (PsiDeclarationStatement)element;
+      final PsiElement grandParent = element.getParent();
+      if (grandParent instanceof PsiForStatement) {
+        splitMultipleDeclarationInForStatementInitialization(declarationStatement);
+        return;
+      }
+      final PsiElement[] elements = declarationStatement.getDeclaredElements();
+      final PsiVariable variable = (PsiVariable)elements[0];
+      variable.normalizeDeclaration();
+      for (int i = 1; i < elements.length; i++) {
+        declarationStatement = PsiTreeUtil.getNextSiblingOfType(declarationStatement, PsiDeclarationStatement.class);
+        assert declarationStatement != null;
+        JavaSharedImplUtil.normalizeBrackets((PsiVariable)declarationStatement.getDeclaredElements()[0]);
+      }
     }
-    final PsiElement greatGrandParent = grandParent.getParent();
-    if (greatGrandParent instanceof PsiForStatement) {
-      final PsiDeclarationStatement declarationStatement =
-        (PsiDeclarationStatement)grandParent;
-      splitMultipleDeclarationInForStatementInitialization(
-        declarationStatement);
-      return;
+    else if (element instanceof PsiField) {
+      PsiField field = DeclarationSearchUtils.findFirstFieldInDeclaration((PsiField)element);
+      PsiField nextField = field;
+      int count = 0;
+      while (nextField != null) {
+        count++;
+        nextField = DeclarationSearchUtils.findNextFieldInDeclaration(nextField);
+      }
+      field.normalizeDeclaration();
+      for (int i = 1; i < count; i++) {
+        field = PsiTreeUtil.getNextSiblingOfType(field, PsiField.class);
+        assert field != null;
+        JavaSharedImplUtil.normalizeBrackets(field);
+      }
     }
-    parent.normalizeDeclaration();
+    else if (element instanceof PsiMethod) {
+      final PsiMethod method = (PsiMethod)element;
+      final PsiTypeElement returnTypeElement = method.getReturnTypeElement();
+      if (returnTypeElement == null) {
+        return;
+      }
+      final PsiType returnType = method.getReturnType();
+      if (returnType == null) {
+        return;
+      }
+      PsiElement child = method.getParameterList();
+      while (!(child instanceof PsiCodeBlock)) {
+        final PsiElement elementToDelete = child;
+        child = child.getNextSibling();
+        if (elementToDelete instanceof PsiJavaToken) {
+          final IElementType tokenType = ((PsiJavaToken)elementToDelete).getTokenType();
+          if (JavaTokenType.LBRACKET.equals(tokenType) || JavaTokenType.RBRACKET.equals(tokenType)) {
+            elementToDelete.delete();
+          }
+        }
+      }
+      final PsiTypeElement typeElement = JavaPsiFacade.getElementFactory(project).createTypeElement(returnType);
+      returnTypeElement.replace(typeElement);
+    }
   }
 
-  private static void splitMultipleDeclarationInForStatementInitialization(
-    PsiDeclarationStatement declarationStatement) {
+  private static void splitMultipleDeclarationInForStatementInitialization(PsiDeclarationStatement declarationStatement) {
     final PsiElement forStatement = declarationStatement.getParent();
-    final PsiElement[] declaredElements =
-      declarationStatement.getDeclaredElements();
+    final PsiElement[] declaredElements = declarationStatement.getDeclaredElements();
     final Project project = forStatement.getProject();
-    final PsiElementFactory factory =
-      JavaPsiFacade.getElementFactory(project);
+    final PsiElementFactory factory = JavaPsiFacade.getElementFactory(project);
     final PsiElement greatGreatGrandParent = forStatement.getParent();
     final PsiBlockStatement blockStatement;
     final PsiCodeBlock codeBlock;
     if (!(greatGreatGrandParent instanceof PsiCodeBlock)) {
-      blockStatement = (PsiBlockStatement)
-        factory.createStatementFromText("{}", forStatement);
+      blockStatement = (PsiBlockStatement)factory.createStatementFromText("{}", forStatement);
       codeBlock = blockStatement.getCodeBlock();
     }
     else {
@@ -85,19 +135,14 @@ public class NormalizeDeclarationFix extends InspectionGadgetsFix {
       final PsiVariable variable = (PsiVariable)declaredElement;
       final PsiType type = variable.getType();
       final String typeText = type.getCanonicalText();
-      final StringBuilder newStatementText =
-        new StringBuilder(typeText);
-      newStatementText.append(' ');
-      newStatementText.append(variable.getName());
+      final StringBuilder newStatementText = new StringBuilder(typeText);
+      newStatementText.append(' ').append(variable.getName());
       final PsiExpression initializer = variable.getInitializer();
       if (initializer != null) {
-        newStatementText.append('=');
-        newStatementText.append(initializer.getText());
+        newStatementText.append('=').append(initializer.getText());
       }
       newStatementText.append(';');
-      final PsiStatement newStatement =
-        factory.createStatementFromText(
-          newStatementText.toString(), forStatement);
+      final PsiStatement newStatement = factory.createStatementFromText(newStatementText.toString(), forStatement);
       if (codeBlock == null) {
         greatGreatGrandParent.addBefore(newStatement, forStatement);
       }
@@ -105,12 +150,17 @@ public class NormalizeDeclarationFix extends InspectionGadgetsFix {
         codeBlock.add(newStatement);
       }
     }
-    for (int i = 1; i < declaredElements.length; i++) {
+    for (int i = 0; i < declaredElements.length; i++) {
       final PsiElement declaredElement = declaredElements[i];
       if (!(declaredElement instanceof PsiVariable)) {
         continue;
       }
-      declaredElement.delete();
+      if (i == 0) {
+        ((PsiVariable)declaredElement).normalizeDeclaration();
+      }
+      else {
+        declaredElement.delete();
+      }
     }
     if (codeBlock != null) {
       codeBlock.add(forStatement);
